service-visualizer/requirements.txt
flask
gunicorn
xgboost==1.0.0
scikit-learn
matplotlib
shap
pyspark
singledispatch >= 3.4.0.3
attrs > 16.0.0
Pillow
